+++ apps/dashboard/src/utils/trpc.ts
@@ -9,7 +9,7 @@ import superjson from "superjson"
 
 const getBaseUrl = () => {
   if (env.NEXT_PUBLIC_NODE_ENV === "production") {
-    return "https://web.online.ntnu.no"
+    return "https://rpc.web.online.ntnu.no"
   }
   return "http://localhost:3002"
 }
