+++ apps/rpc/biome.json
@@ -1,0 +1,4 @@
+{
+  "$schema": "./node_modules/@biomejs/biome/configuration_schema.json",
+  "extends": ["../../biome.json"]
+}
