+++ apps/rpc/package.json
@@ -1,0 +1,30 @@
+{
+  "name": "@dotkomonline/rpc",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "dev": "tsx src/index.ts",
+    "build": "tsc .",
+    "lint": "biome check . --apply",
+    "lint-check": "biome check .",
+    "type-check": "tsc --noEmit"
+  },
+  "dependencies": {
+    "@dotkomonline/env": "workspace:*",
+    "@dotkomonline/gateway-trpc": "workspace:*",
+    "@fastify/cors": "^9.0.1",
+    "@trpc/server": "^10.45.0",
+    "fastify": "^4.28.0",
+    "zod": "^3.22.4"
+  },
+  "devDependencies": {
+    "@biomejs/biome": "1.6.4",
+    "@dotkomonline/config": "workspace:*",
+    "@dotkomonline/tsconfig": "workspace:*",
+    "@types/node": "^20.12.7",
+    "tslib": "^2.6.2",
+    "typescript": "^5.4.5",
+    "tsx": "^4.15.6"
+  }
+}
