+++ apps/rpc/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../packages/tsconfig/tsconfig.json",
+  "include": ["./**/*.ts", "./**/*.tsx"],
+  "exclude": [],
+  "compilerOptions": {
+    "baseUrl": ".",
+    "jsx": "preserve",
+    "incremental": true,
+    "strictNullChecks": true
+  }
+}
