+++ apps/web/src/app/api/trpc/[...path]/route.ts
